# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Cristina Lucero creator site
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  displayName: string;
  username: string;
  bio: string;
  avatarUrl?: string;
  coverUrl?: string;
  location?: string;
  subscriberCount: number;
  totalPhotos: number;
  totalVideos: number;
  joinedDate: string;
}

export interface UpdateProfileBody {
  displayName?: string;
  username?: string;
  bio?: string;
  avatarUrl?: string;
  coverUrl?: string;
  location?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  itemCount: number;
}

export type ContentItemType =
  (typeof ContentItemType)[keyof typeof ContentItemType];

export const ContentItemType = {
  photo: "photo",
  video: "video",
  bundle: "bundle",
} as const;

export interface ContentItem {
  id: number;
  title: string;
  description?: string;
  type: ContentItemType;
  price: number;
  previewUrl: string;
  isLocked: boolean;
  isFeatured: boolean;
  categoryId?: number;
  categoryName?: string;
  likeCount: number;
  commentCount: number;
  createdAt: string;
  tags?: string[];
}

export type CreateContentBodyType =
  (typeof CreateContentBodyType)[keyof typeof CreateContentBodyType];

export const CreateContentBodyType = {
  photo: "photo",
  video: "video",
  bundle: "bundle",
} as const;

export interface CreateContentBody {
  title: string;
  description?: string;
  type: CreateContentBodyType;
  price: number;
  previewUrl: string;
  isLocked: boolean;
  isFeatured: boolean;
  categoryId?: number;
  tags?: string[];
}

export type UpdateContentBodyType =
  (typeof UpdateContentBodyType)[keyof typeof UpdateContentBodyType];

export const UpdateContentBodyType = {
  photo: "photo",
  video: "video",
  bundle: "bundle",
} as const;

export interface UpdateContentBody {
  title?: string;
  description?: string;
  type?: UpdateContentBodyType;
  price?: number;
  previewUrl?: string;
  isLocked?: boolean;
  isFeatured?: boolean;
  categoryId?: number;
  tags?: string[];
}

export interface DeleteResult {
  success: boolean;
  id: number;
}

export interface SiteStats {
  totalContent: number;
  totalPhotos: number;
  totalVideos: number;
  totalBundles: number;
  subscriberCount: number;
}

export interface CreatePurchaseBody {
  contentId: number;
  customerEmail: string;
  customerName: string;
  paymentMethod: string;
}

export interface Purchase {
  id: number;
  contentId: number;
  contentTitle: string;
  customerEmail: string;
  amount: number;
  status: string;
  createdAt: string;
}

export interface VerifyCreatorBody {
  password: string;
}

export interface VerifyCreatorResult {
  success: boolean;
  token: string;
}

export interface CreateInquiryBody {
  name: string;
  email: string;
  whatsapp?: string;
  message: string;
}

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  whatsapp?: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export type ListContentParams = {
  categoryId?: number;
  type?: ListContentType;
};

export type ListContentType =
  (typeof ListContentType)[keyof typeof ListContentType];

export const ListContentType = {
  photo: "photo",
  video: "video",
  bundle: "bundle",
} as const;