# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Cristina Lucero creator site
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inquiries:
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Submit a visitor enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInquiryBody"
      responses:
        "201":
          description: Enquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
        "400":
          description: Invalid request
...
  /creator/inquiries/{id}/read:
    put:
      operationId: markInquiryRead
      tags: [creator]
      summary: Mark an enquiry as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
components:
  schemas:
...
    CreateInquiryBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        whatsapp:
          type: string
        message:
          type: string
      required: [name, email, message]
    Inquiry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        whatsapp:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, message, isRead, createdAt]