# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Cristina Lucero creator site
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get creator profile
 */
export const GetProfileResponse = zod.object({
  id: zod.number(),
  displayName: zod.string(),
  username: zod.string(),
  bio: zod.string(),
  avatarUrl: zod.string().optional(),
  coverUrl: zod.string().optional(),
  location: zod.string().optional(),
  subscriberCount: zod.number(),
  totalPhotos: zod.number(),
  totalVideos: zod.number(),
  joinedDate: zod.string(),
});

/**
 * @summary Update creator profile (creator only)
 */
export const UpdateProfileBody = zod.object({
  displayName: zod.string().optional(),
  username: zod.string().optional(),
  bio: zod.string().optional(),
  avatarUrl: zod.string().optional(),
  coverUrl: zod.string().optional(),
  location: zod.string().optional(),
});

export const UpdateProfileResponse = zod.object({
  id: zod.number(),
  displayName: zod.string(),
  username: zod.string(),
  bio: zod.string(),
  avatarUrl: zod.string().optional(),
  coverUrl: zod.string().optional(),
  location: zod.string().optional(),
  subscriberCount: zod.number(),
  totalPhotos: zod.number(),
  totalVideos: zod.number(),
  joinedDate: zod.string(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  itemCount: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary List all content items
 */
export const ListContentQueryParams = zod.object({
  categoryId: zod.coerce.number().optional(),
  type: zod.enum(["photo", "video", "bundle"]).optional(),
});

export const ListContentResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  type: zod.enum(["photo", "video", "bundle"]),
  price: zod.number(),
  previewUrl: zod.string(),
  isLocked: zod.boolean(),
  isFeatured: zod.boolean(),
  categoryId: zod.number().optional(),
  categoryName: zod.string().optional(),
  likeCount: zod.number(),
  commentCount: zod.number(),
  createdAt: zod.string(),
  tags: zod.array(zod.string()).optional(),
});
export const ListContentResponse = zod.array(ListContentResponseItem);

/**
 * @summary Create a new content item (creator only)
 */
export const CreateContentBody = zod.object({
  title: zod.string(),
  description: zod.string().optional(),
  type: zod.enum(["photo", "video", "bundle"]),
  price: zod.number(),
  previewUrl: zod.string(),
  isLocked: zod.boolean(),
  isFeatured: zod.boolean(),
  categoryId: zod.number().optional(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Get featured content items
 */
export const GetFeaturedContentResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  type: zod.enum(["photo", "video", "bundle"]),
  price: zod.number(),
  previewUrl: zod.string(),
  isLocked: zod.boolean(),
  isFeatured: zod.boolean(),
  categoryId: zod.number().optional(),
  categoryName: zod.string().optional(),
  likeCount: zod.number(),
  commentCount: zod.number(),
  createdAt: zod.string(),
  tags: zod.array(zod.string()).optional(),
});
export const GetFeaturedContentResponse = zod.array(
  GetFeaturedContentResponseItem,
);

/**
 * @summary Get a single content item
 */
export const GetContentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetContentResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  type: zod.enum(["photo", "video", "bundle"]),
  price: zod.number(),
  previewUrl: zod.string(),
  isLocked: zod.boolean(),
  isFeatured: zod.boolean(),
  categoryId: zod.number().optional(),
  categoryName: zod.string().optional(),
  likeCount: zod.number(),
  commentCount: zod.number(),
  createdAt: zod.string(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Update a content item (creator only)
 */
export const UpdateContentParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateContentBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().optional(),
  type: zod.enum(["photo", "video", "bundle"]).optional(),
  price: zod.number().optional(),
  previewUrl: zod.string().optional(),
  isLocked: zod.boolean().optional(),
  isFeatured: zod.boolean().optional(),
  categoryId: zod.number().optional(),
  tags: zod.array(zod.string()).optional(),
});

export const UpdateContentResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  type: zod.enum(["photo", "video", "bundle"]),
  price: zod.number(),
  previewUrl: zod.string(),
  isLocked: zod.boolean(),
  isFeatured: zod.boolean(),
  categoryId: zod.number().optional(),
  categoryName: zod.string().optional(),
  likeCount: zod.number(),
  commentCount: zod.number(),
  createdAt: zod.string(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Delete a content item (creator only)
 */
export const DeleteContentParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteContentResponse = zod.object({
  success: zod.boolean(),
  id: zod.number(),
});

/**
 * @summary Get site statistics
 */
export const GetSiteStatsResponse = zod.object({
  totalContent: zod.number(),
  totalPhotos: zod.number(),
  totalVideos: zod.number(),
  totalBundles: zod.number(),
  subscriberCount: zod.number(),
});

/**
 * @summary List all purchases (creator only)
 */
export const ListPurchasesResponseItem = zod.object({
  id: zod.number(),
  contentId: zod.number(),
  contentTitle: zod.string(),
  customerEmail: zod.string(),
  amount: zod.number(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const ListPurchasesResponse = zod.array(ListPurchasesResponseItem);

/**
 * @summary Create a new purchase
 */
export const CreatePurchaseBody = zod.object({
  contentId: zod.number(),
  customerEmail: zod.string(),
  customerName: zod.string(),
  paymentMethod: zod.string(),
});

/**
 * @summary Verify creator password
 */
export const VerifyCreatorBody = zod.object({
  password: zod.string(),
});

export const VerifyCreatorResponse = zod.object({
  success: zod.boolean(),
  token: zod.string(),
});

/**
 * @summary Submit a visitor enquiry
 */
export const CreateInquiryBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  whatsapp: zod.string().optional(),
  message: zod.string(),
});

/**
 * @summary List all visitor enquiries (creator only)
 */
export const ListInquiriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  whatsapp: zod.string().optional(),
  message: zod.string(),
  isRead: zod.boolean(),
  createdAt: zod.string(),
});
export const ListInquiriesResponse = zod.array(ListInquiriesResponseItem);

/**
 * @summary Mark an enquiry as read
 */
export const MarkInquiryReadParams = zod.object({
  id: zod.coerce.number(),
});

export const MarkInquiryReadResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  whatsapp: zod.string().optional(),
  message: zod.string(),
  isRead: zod.boolean(),
  createdAt: zod.string(),
});